+++ dataverse-webapp/src/main/java/edu/harvard/iq/dataverse/metrics/PublishedFilesChart.java
@@ -68,7 +68,7 @@ public class PublishedFilesChart implements Serializable {
         mode = ChartMode.CUMULATIVE;
 
         if (chartMetrics.isEmpty()) {
-            yearlyStats.add(new ChartMetrics((double) LocalDateTime.now().getYear(), 0L));
+            yearlyStats.add(new ChartMetrics(LocalDateTime.now().getYear(), 0L));
             selectedYear = LocalDate.now().getYear();
         } else {
             yearlyStats = MetricsUtil.countMetricsPerYearAndFillMissingYearsDescending(chartMetrics);
